Fixed Cost Table Populates
Seperated the fixed cost code in to sub-components of FixedCostDetails (FixedCostTable, FixedCostDataInput). In fixedCostDatInput set up refs with type for each element, added v-model on each input, @submit.prevent on the form, and a Add button which calls addFixedCostAction from the reportStore. The new fixed cost objects pushed into the fixedCosts state array populate in the FixedCostTable.