hotfix duplicate fixed cost components (#7)
* chore: add ids to fixed costs

* chore: update project layout and structure

* installed Vue-Router

* Fixed Cost Table Populates

Seperated the fixed cost code in to sub-components of FixedCostDetails (FixedCostTable, FixedCostDataInput). In fixedCostDatInput set up refs with type for each element, added v-model on each input, @submit.prevent on the form, and a Add button which calls addFixedCostAction from the reportStore. The new fixed cost objects pushed into the fixedCosts state array populate in the FixedCostTable.

* Used Prettier on code. Added more types.ts files & imported

* Duplicate gone (somehow) commit to compare with broken branch

* Confirming hotfix works

---------

Co-authored-by: Dan Andrade <dan@cloudspot.io>