feature vue router (#10)
* added store/stepStore.ts

* Actions on store and component

Added actions to stepStore.ts (for incrementing and decrementing the step state). Attached actions with v-on (@) to buttons in CobdFooter)

* Imported stepsArray into stepStore.js

* stepName updates

Created a gettin in stepStore.ts 'updateStep', returned the state.stepName = stepsArray[state.stepName].path and called the getting in each action (forward..., back....). Had to set the type on the updateState getter to 'any'

* Added views/Home.vue as step 0

* fixed router for home link

In stepsArray changed Home path from '' to '/'. In routes/index.ts set the if condition to step.path === '/#'. Now works as expected. Clicking home brings to home.

* Rouuting works with error

setting the 'to' on router link as ':to='stepStore.stepName'' (notice : before to) allows routing to the current state of stepName to work. BUT getting error 'Uncaught TypeError: this.updateStep is not a function' when the actions are used (fowardStepAction, backStepAction)

* Error fixed

Removed getter and put logic in actions to stop error

* set min/max for stepNum

using conitional inside stepStore actions (back.... forward...) set min/max value of this.stepNum constrain within stepsArray

* Clean up

Removed comments, removed unneeded arguments from actions, code formatting