refactor form reusability (#12)
* Commit to checkout development

* removed extra FixedCostsStep file in FixedCostComponents

* Fixed cost refactor, On FinalReportStep

Moved the data binding (v-for, key) from the parent step into FixedCostsTable to reduce duplication on imports and preserve styling. Imported FixedCostsTable into FinalReportStep

* Changing task to building a reusuable form, will delete this branch

* Data wont pass to Pinia

Once the form element is abstracted in it's own file I can't get the data to flow through to pinia store

* Commit to save

* DataInput works with error

DataInput now passes data through FixedCostData to Pinia store, but still recieving console error: '[Vue warn]: Component emitted event update:modelValue but it is neither declared in the emits option nor as an onUpdate:modelValue prop.'

* Abstraced option works

Needed to v-bind value on the option (i.e. :value='' instead of value='') to get it to pass data through.

* Created reusable Select

Created a DataSelect component that can be reused across forms. Sends option to the pinia store and to the form from there.

* Removed console logs, comments, unused imports