feature costs table (#19)
* Styling data table

* Tailwind not updating classes

* Formated as table with issues

Data lays out in a table format via flexbox. Howeve, still having issues with tailwind not updating the classnames, etc

* Building fixed cost total calulation

the fixed cost amount needs to be multiplied by the pay period. Figuring out how to convert the pay period string into the approprate number

* Changed name of main.css to output.css. Commiting to create sub branch to attempt a Tailwind fix

* Indivdual Fixed Cost calculates based on pay period

in fixedCostStore -> addFixedCostAction I find an element which eqauls the payPeriod, then set the multiplier to equal the multiplier in the element.

* Reformat, removed console.logs and comments

* Variable costs now total