refactor report fields (#32)
* set up userEmail validation, disabled button if not valid

* Does not validate if user does not interact

the field does not validate if the user does not click in, so even if it is required and left blank the validation does not trigger. It does work if the user clicks in and enters something, then deletes it.

* end of day

* Commit to switch branch

* UserEmailStep Validates

initialized the email.value to empty string and updateInputValidAction to false within onMounted. This validates the form automatically when the DOM mounts so the user cannot skip the step

* Commit to save

* companyName step validates

* End of day

* validation works, value remains in input on previous, but input does not revalidate

* using onUpdate form input revalidates via updateInputValidAction when user changes data

* end of day

* Cannot figure out how to run the first validation on Next click

* Commit to switch branch

* styled home button

* refactored computed get/set

* Built array, attempting to destructure with v-for

* Used v-for and slots to reduce final report step down to a single component. All logic is in reportOutputsArray

* added key field to reportOutputs