feature vee validate (#34)
* set up userEmail validation, disabled button if not valid

* Does not validate if user does not interact

the field does not validate if the user does not click in, so even if it is required and left blank the validation does not trigger. It does work if the user clicks in and enters something, then deletes it.

* end of day

* Commit to switch branch

* UserEmailStep Validates

initialized the email.value to empty string and updateInputValidAction to false within onMounted. This validates the form automatically when the DOM mounts so the user cannot skip the step

* Commit to save

* companyName step validates

* End of day

* validation works, value remains in input on previous, but input does not revalidate

* using onUpdate form input revalidates via updateInputValidAction when user changes data

* end of day

* Cannot figure out how to run the first validation on Next click

* Commit to switch branch

* styled home button

* refactored computed get/set

* refactoring how forward/backward step actions work

* adjusted routing to go off of next/previous in stepsArray

* end of day

* Footer buttons conditionally render

Fixed the getIndex function in stepStore (was not formated correctlt) and fixed the conditional which sets this.stepNum (was returning a value not setting stepNum=) which allowed the stepNum to update properly in CobdFooter.

* Trying to validate FixedCostData...

* Forms validate on Fixed Cost Step, error message displays, need to validate on click not on blur

* Reverted final report back to before it was reduced

* CompanyName and UserEmail now don't allow user to proceed if value is deleted (asyc await and resp.valid added)

* Researching Provide/Inject next

* Added back the shorted report-output with reportOUtputsArray in the FinalReportStep

* Add onMount again for CompanyNameStep, will merge to fix development branch