refactor handler functions (#57)
* fixedCost Delete handler refactored

* variableCost delete handler removed

* commit due to forgetting to save

* removed fixedCostEditHandler, moved the work into openEditFormModal() in modal store, added editFixedCostActionWrappter to open/close necessary modals

* removed wrapper, now all necessary actions are within openEditFormModal

* cleared error '[Vue warn]: Extraneous non-props attributes (class) were passed to component but could not be automatically inherited because component renders fragment or text root nodes.'

cannot have an attribute (e.g. class) on a resusable component named the same as one used on the component call. So either rename the attribute in the component or do not use it on the import

* refactored variablecosts removing handlers

had to add a type to openEditFormModal and seletedCostAction (also an if statement here to find id from the correct cost array)

* clean up

deleted console.logs and the handleEditCost function file